requirements.txt
Django==1.8.2
Pillow==2.8.1
django-bootstrap3-datetimepicker==2.2.3
django-datatable-view==0.8.1
django-input-mask==1.3.7.1
django-localflavor==1.1
django-tinymce==1.5.3
python-dateutil==2.4.2
six==1.9.0
sorl-thumbnail==12.2
